python/microsoft_agents_m365copilot_core/tests/authentication/test_azure_identity_authentication_provider.py
from azure.identity import EnvironmentCredential
from kiota_abstractions.authentication import AuthenticationProvider

from microsoft_agents_m365copilot_core.authentication import AzureIdentityAuthenticationProvider


def test_subclassing():
    assert issubclass(AzureIdentityAuthenticationProvider, AuthenticationProvider)
